+++ table/result/indexed/indexed.go
@@ -1,0 +1,5 @@
+package indexed
+
+// Value is a type scan destination of ydb values
+// This is a proxy type for preparing go1.18 embedded types
+type Value interface{}